// src/app/ficha-de-matricula/models/matricula.ts
export class DatosDeAlumno {
    matriculaId: number = 0;
    fotografia: string = "";
    fotografiaArchivo: string = "";
    nombre: string = "";
    apellidoPaterno: string = "";
    apellidoMaterno: string = "";
    generoId: number = 0;
    nombreSocial: string = "";
    fechaDeNacimiento: string = "";
    paisId: number = 0;
    etniaId: number = 0;
    religionId: number = 0;
    calle: string = "";
    numero: number = 0;
    departamento: string = "";
    regionId: number = 0;
    comunaId: number = 0;
    celular: string = "";
    correoElectronico: string = "";
    viveCon: string = "";
    hermanosEnColegio: string[] = [];
}